Fix a crash on big-endian platforms when sizeof(int) != sizeof(long).

+++ src/place.c
@@ -409,8 +409,8 @@ void CenterClient(const BoundingBox *box, ClientNode *np)
 /** Compare two integers. */
 int IntComparator(const void *a, const void *b)
 {
-   const int ia = *(const long*)a;
-   const int ib = *(const long*)b;
+   const int ia = *(const int*)a;
+   const int ib = *(const int*)b;
    return ia - ib;
 }
